@@ -30,16 +30,16 @@ class UserSecurity extends CustomObject
 		$this->export = Input::convertToBool($this->export);
 		$this->changeSettings = Input::convertToBool($this->changeSettings);
 		$this->locked = Input::convertToBool($this->locked);
+	}
 
+	public function link()
+	{
 		$this->viewIcon = Icon::load($this->view ? 'check' : 'x');
 		$this->editIcon = Icon::load($this->edit ? 'check' : 'x');
 		$this->exportIcon = Icon::load($this->export ? 'check' : 'x');
 		$this->changeSettingsIcon = Icon::load($this->changeSettings ? 'check' : 'x');
 		$this->lockedIcon = Icon::load($this->locked ? 'check' : 'x');
-	}
 
-	public function link()
-	{
 		$this->module = (new Module)->get($this->moduleId)[0];
 		$this->user = (new LocalUser)->get($this->userId)[0];
 
